Amiteh_Back1.2.1/pyproject.toml
[tool.poetry]
name = "nordichomes"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ruff = "^0.4.2"
pyright = "^1.1.361"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
